+++ skforecast/direct/_forecaster_direct_multivariate.py
@@ -54,6 +54,7 @@ from ..utils import (
     manage_warnings,
     get_style_repr_html,
     _build_predict_function,
+    scale_correction_factor_differentiation
 )
 from ..preprocessing import TimeSeriesDifferentiator, QuantileBinner
 
@@ -2454,7 +2455,7 @@ class ForecasterDirectMultiVariate(ForecasterBase):
 
         predictions = self._direct_predict(steps=steps, Xs=Xs)
 
-        if self.differentiation is not None:
+        if differentiator is not None:
             predictions = differentiator.inverse_transform_next_window(predictions)
         
         predictions = transform_numpy(
@@ -2592,7 +2593,7 @@ class ForecasterDirectMultiVariate(ForecasterBase):
         boot_columns = [f"pred_boot_{i}" for i in range(n_boot)]
         boot_predictions = boot_predictions + sampled_residuals
 
-        if self.differentiation is not None:
+        if differentiator is not None:
             boot_predictions = (
                 differentiator.inverse_transform_next_window(boot_predictions)
             )
@@ -2713,22 +2714,13 @@ class ForecasterDirectMultiVariate(ForecasterBase):
         else:
             correction_factor = np.quantile(np.abs(residuals), nominal_coverage)
 
-        if self.differentiation is not None:
+        if differentiator is not None:
             predictions = differentiator.inverse_transform_next_window(predictions)
-            
-            # Since the inverse transform is a cumsum, the prediction interval needs to be 
-            # scaled according to the differentiation order. For d=1, the variance grows 
-            # linearly with h, so the interval grows with sqrt(h). For d > 1, we use the 
-            # square root of the accumulated variances of the MA(inf) representation.
-            import scipy.special
-            d = self.differentiator_[self.level].order
-            steps_array = np.arange(1, len(predictions) + 1)
-            # The scaling factor is exactly the square root of the accumulated variances 
-            # of the MA(infinity) representation of the (1-B)^{-d} filter
-            scaling_factor = np.sqrt(
-                np.cumsum(scipy.special.comb(steps_array + d - 2, d - 1)**2)
+            correction_factor = scale_correction_factor_differentiation(
+                correction_factor     = correction_factor,
+                steps                 = len(predictions),
+                differentiation_order = differentiator.order
             )
-            correction_factor = correction_factor * scaling_factor
 
         lower_bound = predictions - correction_factor
         upper_bound = predictions + correction_factor