+++ skforecast/recursive/_forecaster_recursive.py
@@ -47,7 +47,8 @@ from ..utils import (
     get_style_repr_html,
     set_cpu_gpu_device,
     _build_predict_function,
-    manage_warnings    
+    manage_warnings,
+    scale_correction_factor_differentiation
 )
 from ..preprocessing import TimeSeriesDifferentiator, QuantileBinner
 
@@ -2237,20 +2238,11 @@ class ForecasterRecursive(ForecasterBase):
 
         if differentiator is not None:
             predictions = differentiator.inverse_transform_next_window(predictions)
-            
-            # Since the inverse transform is a cumsum, the prediction interval needs to be 
-            # scaled according to the differentiation order. For d=1, the variance grows 
-            # linearly with h, so the interval grows with sqrt(h). For d > 1, we use the 
-            # square root of the accumulated variances of the MA(inf) representation.
-            import scipy.special
-            d = self.differentiation
-            steps_array = np.arange(1, len(predictions) + 1)
-            # The scaling factor is exactly the square root of the accumulated variances 
-            # of the MA(infinity) representation of the (1-B)^{-d} filter
-            scaling_factor = np.sqrt(
-                np.cumsum(scipy.special.comb(steps_array + d - 2, d - 1)**2)
+            correction_factor = scale_correction_factor_differentiation(
+                correction_factor     = correction_factor,
+                steps                 = len(predictions),
+                differentiation_order = self.differentiation
             )
-            correction_factor = correction_factor * scaling_factor
 
         lower_bound = predictions - correction_factor
         upper_bound = predictions + correction_factor